moved game into flat directory structure #4
