+++ cmd/serveCommand/serveCommand.go
@@ -5,8 +5,8 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/UnnecessaryRain/ironway-core/pkg/game"
-	"github.com/UnnecessaryRain/ironway-core/pkg/interpreter"
+	"github.com/UnnecessaryRain/ironway-core/pkg/mud/game"
+	"github.com/UnnecessaryRain/ironway-core/pkg/mud/interpreter"
 	"github.com/UnnecessaryRain/ironway-core/pkg/network/client"
 
 	"github.com/UnnecessaryRain/ironway-core/pkg/network/server"
